+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Fix: Columns layout for currency table.
 - Fix: Error when updating day or night landings in a flight record.
 - Update: Added recalculation of night time during flight record import.
 - Fix: Corrected night time calculation for flights with identical departure and arrival locations (e.g., training circuits).